Refactored code to use bedrock mixin
