+++ controlpanel/api/models/user.py
@@ -3,6 +3,7 @@ from crequest.middleware import CrequestMiddleware
 from django.contrib.auth.models import AbstractUser
 from django.contrib.auth.signals import user_logged_in
 from django.db import models
+from django.urls import reverse
 from django.utils.functional import cached_property
 
 # First-party/Local
@@ -191,6 +192,9 @@ class User(AbstractUser):
                     "migration_state",
                 ],
             )
+
+    def get_absolute_url(self):
+        return reverse("manage-user", kwargs={"pk": self.auth0_id})
 
 
 user_logged_in.connect(prometheus_login_event)
